+++ windows/flutter/generated_plugin_registrant.cc
@@ -8,6 +8,7 @@
 #include <flutter_js/flutter_js_plugin.h>
 #include <flutter_webview/flutter_webview_plugin.h>
 #include <url_launcher_windows/url_launcher_plugin.h>
+#include <windows_speak/windows_speak_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
   FileChooserPluginRegisterWithRegistrar(
@@ -18,4 +19,6 @@ void RegisterPlugins(flutter::PluginRegistry* registry) {
       registry->GetRegistrarForPlugin("FlutterWebviewPlugin"));
   UrlLauncherPluginRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("UrlLauncherPlugin"));
+  WindowsSpeakPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("WindowsSpeakPlugin"));
 }